requirements.txt
Jinja2==3.1.2
itsdangerous==2.1.2
werkzeug==3.0.1
Flask==3.0.0
flask_openapi3==3.0.1
Flask-SQLAlchemy==3.1.1
pyodbc==5.0.1
pytest==7.4.4
